backend/requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-firestore
google-cloud-aiplatform
google-cloud-bigquery
pyarrow
pydantic
pydantic-settings
pandas
openpyxl
fuzzywuzzy
python-levenshtein
python-multipart
vertexai
google-generativeai
firebase-admin
firebase-functions
python-dateutil
tabulate
reportlab
weasyprint
jinja2
db-dtypes
PyYAML==6.0.1